Refactor: 역할에 따른 모듈 분리
